requirements.txt
requests
python-dotenv
openai
watchdog
